refactor: give the language list its own module
The language table was the only thing in datamodels that pulled in Qt,
and only for the translation markers. Everything else there is plain
data the rest of the app passes around, so the list never really fit.
