+++ mpvqc/datamodels.py
@@ -6,8 +6,6 @@ from dataclasses import dataclass
 from enum import Enum, IntEnum, auto
 from pathlib import Path
 
-from PySide6.QtCore import QT_TRANSLATE_NOOP
-
 
 @dataclass(frozen=True, slots=True)
 class Comment:
@@ -55,42 +53,3 @@ class DocumentImportResult:
     existing_videos: tuple[Path, ...]
     existing_subtitles: tuple[Path, ...]
     comments: tuple[Comment, ...]
-
-
-@dataclass(frozen=True)
-class Language:
-    language: str
-    identifier: str
-    translators: tuple[str, ...] = ()
-
-
-LANGUAGES = (
-    Language(
-        language=str(QT_TRANSLATE_NOOP("Languages", "German")),
-        identifier="de-DE",
-    ),
-    Language(
-        language=str(QT_TRANSLATE_NOOP("Languages", "English")),
-        identifier="en-US",
-    ),
-    Language(
-        language=str(QT_TRANSLATE_NOOP("Languages", "Spanish")),
-        identifier="es-MX",
-        translators=("CiferrC",),
-    ),
-    Language(
-        language=str(QT_TRANSLATE_NOOP("Languages", "Hebrew")),
-        identifier="he-IL",
-        translators=("cN3rd",),
-    ),
-    Language(
-        language=str(QT_TRANSLATE_NOOP("Languages", "Italian")),
-        identifier="it-IT",
-        translators=("maddo",),
-    ),
-    Language(
-        language=str(QT_TRANSLATE_NOOP("Languages", "Portuguese")),
-        identifier="pt-PT",
-        translators=("Diogo_23",),
-    ),
-)
